Move lancer initiative appearance form
